// template/react-starter-kit/src/.generated/api/model/inquiryMessageItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { InquiryMessageAuthorRole } from './inquiryMessageAuthorRole';

export interface InquiryMessageItemDto {
  id: string;
  inquiryId: string;
  authorId: string;
  authorName: string;
  authorRole: InquiryMessageAuthorRole;
  content: string;
  createdAt: string;
}